requirements.txt
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
numpy==1.24.3
joblib==1.3.2
pdf2image==1.16.3
pytesseract==0.3.10
pillow==10.0.0
transformers==4.30.2
spacy==3.6.1
nltk==3.8.1
setuptools>=65.5.0